Move UserRegistrationRollbackService auth -> user
