+++ src/main/java/site/hirecruit/hr/domain/auth/aop/UserSessionInfoUpdateAspect.kt
@@ -26,7 +26,7 @@ class UserSessionInfoUpdateAspect (
     @Pointcut("execution(* site.hirecruit.hr.domain.auth.service.UserAuthService+.authentication(..))")
     private fun userAuthService_AuthenticationMethodPointCut(){}
 
-    @Pointcut("execution(* site.hirecruit.hr.domain.auth.service.UserRegistrationRollbackService+.rollback(..))")
+    @Pointcut("execution(* site.hirecruit.hr.domain.user.service.UserRegistrationRollbackService+.rollback(..))")
     private fun userRegistrationRollbackService_rollback(){}
 
     @Pointcut("execution(* site.hirecruit.hr.domain.user.service.UserUpdateService+.update(..))")
