+++ src/test/java/site/hirecruit/hr/domain/auth/aop/UserSessionInfoUpdateAspectTest.kt
@@ -12,7 +12,7 @@ import site.hirecruit.hr.domain.auth.dto.AuthUserInfo
 import site.hirecruit.hr.domain.auth.dto.OAuthAttributes
 import site.hirecruit.hr.domain.user.entity.Role
 import site.hirecruit.hr.domain.auth.service.UserAuthService
-import site.hirecruit.hr.domain.auth.service.UserRegistrationRollbackService
+import site.hirecruit.hr.domain.user.service.UserRegistrationRollbackService
 import site.hirecruit.hr.global.data.SessionAttribute
 import kotlin.random.Random
